Create README.md
After analyzing the gaming data for Heroes Of Pymoli we can see that the majority of the user base is primarily male dominated  at 84%. With the majority age group being young adults in the age range of 20-24 years of age.  The males on average also spent 5x more than the females, with Final Critic being the most purchased item.